+++ plugins/woocommerce/includes/gateways/paypal/class-wc-gateway-paypal.php
@@ -16,6 +16,7 @@ use Automattic\WooCommerce\Gateways\PayPal\Buttons as PayPalButtons;
 use Automattic\WooCommerce\Gateways\PayPal\Helper as PayPalHelper;
 use Automattic\WooCommerce\Gateways\PayPal\Notices as PayPalNotices;
 use Automattic\WooCommerce\Gateways\PayPal\TransactAccountManager as PayPalTransactAccountManager;
+use Automattic\WooCommerce\Gateways\PayPal\Request as PayPalRequest;
 use Automattic\Jetpack\Connection\Manager as Jetpack_Connection_Manager;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -245,8 +246,7 @@ class WC_Gateway_Paypal extends WC_Payment_Gateway {
 		}
 
 		try {
-			include_once WC_ABSPATH . 'includes/gateways/paypal/includes/class-wc-gateway-paypal-request.php';
-			$paypal_request       = new WC_Gateway_Paypal_Request( $this );
+			$paypal_request       = new PayPalRequest( $this );
 			$paypal_order_details = $paypal_request->get_paypal_order_details( $paypal_order_id );
 
 			// Update the shipping information.
@@ -592,11 +592,11 @@ class WC_Gateway_Paypal extends WC_Payment_Gateway {
 	public function process_payment( $order_id ) {
 		include_once __DIR__ . '/includes/class-wc-gateway-paypal-request.php';
 
-		$order          = wc_get_order( $order_id );
-		$paypal_request = new WC_Gateway_Paypal_Request( $this );
+		$order = wc_get_order( $order_id );
 
 		if ( $this->should_use_orders_v2() ) {
-			$paypal_order = $paypal_request->create_paypal_order( $order );
+			$paypal_request = new PayPalRequest( $this );
+			$paypal_order   = $paypal_request->create_paypal_order( $order );
 			if ( ! $paypal_order || empty( $paypal_order['id'] ) || empty( $paypal_order['redirect_url'] ) ) {
 				throw new Exception(
 					esc_html__( 'We are unable to process your PayPal payment at this time. Please try again or use a different payment method.', 'woocommerce' )
@@ -605,7 +605,8 @@ class WC_Gateway_Paypal extends WC_Payment_Gateway {
 
 			$redirect_url = $paypal_order['redirect_url'];
 		} else {
-			$redirect_url = $paypal_request->get_request_url( $order, $this->testmode );
+			$paypal_request = new WC_Gateway_Paypal_Request( $this );
+			$redirect_url   = $paypal_request->get_request_url( $order, $this->testmode );
 		}
 
 		return array(
@@ -692,9 +693,7 @@ class WC_Gateway_Paypal extends WC_Payment_Gateway {
 		$order = wc_get_order( $order_id );
 
 		if ( $this->should_use_orders_v2() ) {
-			include_once __DIR__ . '/includes/class-wc-gateway-paypal-request.php';
-
-			$paypal_request = new WC_Gateway_Paypal_Request( $this );
+			$paypal_request = new PayPalRequest( $this );
 			$paypal_request->capture_authorized_payment( $order );
 			return;
 		}
