+++ plugins/woocommerce/includes/rest-api/Controllers/Version3/class-wc-rest-paypal-buttons-controller.php
@@ -13,10 +13,7 @@ defined( 'ABSPATH' ) || exit;
 
 use Automattic\WooCommerce\Enums\OrderStatus;
 use Automattic\WooCommerce\Gateways\PayPal\Constants as PayPalConstants;
-
-if ( ! class_exists( 'WC_Gateway_Paypal_Request' ) ) {
-	require_once WC_ABSPATH . 'includes/gateways/paypal/includes/class-wc-gateway-paypal-request.php';
-}
+use Automattic\WooCommerce\Gateways\PayPal\Request as PayPalRequest;
 
 /**
  * REST API PayPal buttons controller class.
@@ -135,7 +132,7 @@ class WC_REST_Paypal_Buttons_Controller extends WC_REST_Controller {
 		$order->set_payment_method( $gateway->id );
 		$order->save();
 
-		$paypal_request = new WC_Gateway_Paypal_Request( $gateway );
+		$paypal_request = new PayPalRequest( $gateway );
 		$paypal_order   = $paypal_request->create_paypal_order(
 			$order,
 			$payment_source,
