+++ plugins/woocommerce/src/Gateways/PayPal/Buttons.php
@@ -10,15 +10,11 @@ declare(strict_types=1);
 namespace Automattic\WooCommerce\Gateways\PayPal;
 
 use Automattic\WooCommerce\Proxies\LegacyProxy;
+use Automattic\WooCommerce\Gateways\PayPal\Request as PayPalRequest;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
-
-if ( ! class_exists( 'WC_Gateway_Paypal_Request' ) ) {
-	require_once WC_ABSPATH . 'includes/gateways/paypal/includes/class-wc-gateway-paypal-request.php';
-}
-
 /**
  * Handles PayPal Buttons.
  */
@@ -47,7 +43,7 @@ class Buttons {
 	/**
 	 * The request instance.
 	 *
-	 * @var \WC_Gateway_Paypal_Request
+	 * @var PayPalRequest
 	 */
 	private $request;
 
@@ -58,7 +54,7 @@ class Buttons {
 	 */
 	public function __construct( \WC_Gateway_Paypal $gateway ) {
 		$this->gateway = $gateway;
-		$this->request = new \WC_Gateway_Paypal_Request( $this->gateway );
+		$this->request = new PayPalRequest( $this->gateway );
 
 		// phpcs:ignore Generic.Commenting.Todo.TaskFound
 		$this->enabled = $this->gateway->should_use_orders_v2() && 'yes' === $this->gateway->get_option( 'paypal_buttons', 'yes' );
