+++ plugins/woocommerce/src/Gateways/PayPal/WebhookHandler.php
@@ -11,16 +11,12 @@ namespace Automattic\WooCommerce\Gateways\PayPal;
 
 use Automattic\WooCommerce\Gateways\PayPal\Constants as PayPalConstants;
 use Automattic\WooCommerce\Gateways\PayPal\Helper as PayPalHelper;
+use Automattic\WooCommerce\Gateways\PayPal\Request as PayPalRequest;
 use Automattic\WooCommerce\Enums\OrderStatus;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
-
-if ( ! class_exists( 'WC_Gateway_Paypal_Request' ) ) {
-	require_once WC_ABSPATH . 'includes/gateways/paypal/includes/class-wc-gateway-paypal-request.php';
-}
-
 /**
  * Handles webhook events.
  */
@@ -206,8 +202,8 @@ class WebhookHandler {
 	 * Capture the payment.
 	 *
 	 * @param \WC_Order $order The order object.
-	 * @param array    $links The links from the webhook event.
-	 * @param string   $action The action to perform (capture or authorize).
+	 * @param array     $links The links from the webhook event.
+	 * @param string    $action The action to perform (capture or authorize).
 	 * @return void
 	 */
 	private function authorize_or_capture_payment( $order, $links, $action ) {
@@ -219,7 +215,7 @@ class WebhookHandler {
 			return;
 		}
 		$gateway        = $payment_gateways['paypal'];
-		$paypal_request = new \WC_Gateway_Paypal_Request( $gateway );
+		$paypal_request = new PayPalRequest( $gateway );
 		$paypal_request->authorize_or_capture_payment( $order, $action_url, $action );
 	}
